Clear up all the Dialogue*Service classes
