+++ conjure-java-core/src/test/resources/test/api/BlockingCookieService.java.dialogue
@@ -1,9 +1,24 @@
 package test.api;
 
+import com.palantir.dialogue.Channel;
+import com.palantir.dialogue.ConjureRuntime;
+import com.palantir.dialogue.RemoteExceptions;
 import com.palantir.tokens.auth.BearerToken;
+import java.lang.Override;
 import javax.annotation.Generated;
 
 @Generated("com.palantir.conjure.java.services.dialogue.DialogueServiceGenerator")
 public interface BlockingCookieService {
     void eatCookies(BearerToken token);
+
+    /** Creates a synchronous/blocking client for a CookieService service. */
+    static BlockingCookieService of(Channel channel, ConjureRuntime runtime) {
+        AsyncCookieService delegate = AsyncCookieService.of(channel, runtime);
+        return new BlockingCookieService() {
+            @Override
+            public void eatCookies(BearerToken token) {
+                RemoteExceptions.getUnchecked(delegate.eatCookies(token));
+            }
+        };
+    }
 }
