+++ conjure-java-core/src/test/resources/test/api/BlockingTestService.java.dialogue.prefix
@@ -1,10 +1,14 @@
 package test.prefix.com.palantir.another;
 
 import com.palantir.dialogue.BinaryRequestBody;
+import com.palantir.dialogue.Channel;
+import com.palantir.dialogue.ConjureRuntime;
+import com.palantir.dialogue.RemoteExceptions;
 import com.palantir.ri.ResourceIdentifier;
 import com.palantir.tokens.auth.AuthHeader;
 import java.io.InputStream;
 import java.lang.Deprecated;
+import java.lang.Override;
 import java.lang.String;
 import java.nio.ByteBuffer;
 import java.util.Map;
@@ -90,4 +94,168 @@ public interface BlockingTestService {
 
     void getForStrings(
             AuthHeader authHeader, ResourceIdentifier datasetRid, Set<AliasedString> strings);
+
+    /** Creates a synchronous/blocking client for a TestService service. */
+    static BlockingTestService of(Channel channel, ConjureRuntime runtime) {
+        AsyncTestService delegate = AsyncTestService.of(channel, runtime);
+        return new BlockingTestService() {
+            @Override
+            public Map<String, BackingFileSystem> getFileSystems(AuthHeader authHeader) {
+                return RemoteExceptions.getUnchecked(delegate.getFileSystems(authHeader));
+            }
+
+            @Override
+            public Dataset createDataset(
+                    AuthHeader authHeader, String testHeaderArg, CreateDatasetRequest request) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.createDataset(authHeader, testHeaderArg, request));
+            }
+
+            @Override
+            public Optional<Dataset> getDataset(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(delegate.getDataset(authHeader, datasetRid));
+            }
+
+            @Override
+            public InputStream getRawData(AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(delegate.getRawData(authHeader, datasetRid));
+            }
+
+            @Override
+            public NestedAliasedBinary getAliasedRawData(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.getAliasedRawData(authHeader, datasetRid));
+            }
+
+            @Override
+            public Optional<ByteBuffer> maybeGetRawData(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.maybeGetRawData(authHeader, datasetRid));
+            }
+
+            @Override
+            public AliasedString getAliasedString(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.getAliasedString(authHeader, datasetRid));
+            }
+
+            @Override
+            public void uploadRawData(AuthHeader authHeader, BinaryRequestBody input) {
+                RemoteExceptions.getUnchecked(delegate.uploadRawData(authHeader, input));
+            }
+
+            @Override
+            public void uploadAliasedRawData(AuthHeader authHeader, NestedAliasedBinary input) {
+                RemoteExceptions.getUnchecked(delegate.uploadAliasedRawData(authHeader, input));
+            }
+
+            @Override
+            public Set<String> getBranches(AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(delegate.getBranches(authHeader, datasetRid));
+            }
+
+            @Override
+            public Set<String> getBranchesDeprecated(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.getBranchesDeprecated(authHeader, datasetRid));
+            }
+
+            @Override
+            public Optional<String> resolveBranch(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid, String branch) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.resolveBranch(authHeader, datasetRid, branch));
+            }
+
+            @Override
+            public Optional<String> testParam(
+                    AuthHeader authHeader, ResourceIdentifier datasetRid) {
+                return RemoteExceptions.getUnchecked(delegate.testParam(authHeader, datasetRid));
+            }
+
+            @Override
+            public int testQueryParams(
+                    AuthHeader authHeader,
+                    ResourceIdentifier something,
+                    ResourceIdentifier implicit,
+                    Optional<ResourceIdentifier> optionalMiddle,
+                    Set<String> setEnd,
+                    Optional<ResourceIdentifier> optionalEnd,
+                    String query) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.testQueryParams(
+                                authHeader,
+                                something,
+                                implicit,
+                                optionalMiddle,
+                                setEnd,
+                                optionalEnd,
+                                query));
+            }
+
+            @Override
+            public void testNoResponseQueryParams(
+                    AuthHeader authHeader,
+                    ResourceIdentifier something,
+                    ResourceIdentifier implicit,
+                    Optional<ResourceIdentifier> optionalMiddle,
+                    Set<String> setEnd,
+                    Optional<ResourceIdentifier> optionalEnd,
+                    String query) {
+                RemoteExceptions.getUnchecked(
+                        delegate.testNoResponseQueryParams(
+                                authHeader,
+                                something,
+                                implicit,
+                                optionalMiddle,
+                                setEnd,
+                                optionalEnd,
+                                query));
+            }
+
+            @Override
+            public boolean testBoolean(AuthHeader authHeader) {
+                return RemoteExceptions.getUnchecked(delegate.testBoolean(authHeader));
+            }
+
+            @Override
+            public double testDouble(AuthHeader authHeader) {
+                return RemoteExceptions.getUnchecked(delegate.testDouble(authHeader));
+            }
+
+            @Override
+            public int testInteger(AuthHeader authHeader) {
+                return RemoteExceptions.getUnchecked(delegate.testInteger(authHeader));
+            }
+
+            @Override
+            public Optional<String> testPostOptional(
+                    AuthHeader authHeader, Optional<String> maybeString) {
+                return RemoteExceptions.getUnchecked(
+                        delegate.testPostOptional(authHeader, maybeString));
+            }
+
+            @Override
+            public void testOptionalIntegerAndDouble(
+                    AuthHeader authHeader, OptionalInt maybeInteger, OptionalDouble maybeDouble) {
+                RemoteExceptions.getUnchecked(
+                        delegate.testOptionalIntegerAndDouble(
+                                authHeader, maybeInteger, maybeDouble));
+            }
+
+            @Override
+            public void getForStrings(
+                    AuthHeader authHeader,
+                    ResourceIdentifier datasetRid,
+                    Set<AliasedString> strings) {
+                RemoteExceptions.getUnchecked(
+                        delegate.getForStrings(authHeader, datasetRid, strings));
+            }
+        };
+    }
 }
